requirements.txt
# Backend Framework
fastapi==0.109.0
uvicorn==0.24.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Payment Integration
stripe==9.1.0

# AI & ML
openai==1.12.0
transformers==4.37.2
torch==2.2.0

# Utilities
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.1.0

# Testing
pytest==8.0.0
pytest-cov==4.1.0
httpx==0.26.0

# Deployment
gunicorn==21.2.0
